refactor: ♻️ refator filters
