+++ app/src/components/classes/list/client.tsx
@@ -4,10 +4,10 @@ import { cn } from '@pedaki/design/utils';
 import type { Field } from '@pedaki/services/classes/query.model';
 import type { ClassColumnDef, ClassData } from '~/components/classes/list/columns.tsx';
 import { generateColumns } from '~/components/classes/list/columns.tsx';
-import Filters from '~/components/classes/list/filters.tsx';
 import { searchParams, serialize } from '~/components/classes/list/parameters.ts';
 import { ColumnSelector } from '~/components/datatable/column-selector';
 import { DataTable } from '~/components/datatable/data-table';
+import Filters from '~/components/datatable/filters';
 import Footer from '~/components/datatable/footer';
 import { useScopedI18n } from '~/locales/client.ts';
 import { api } from '~/server/clients/client.ts';
@@ -125,7 +125,7 @@ const Client = ({ className }: React.HTMLAttributes<HTMLDivElement>) => {
   return (
     <div className={cn('flex h-full flex-col gap-4', className)}>
       <div className="flex gap-4">
-        <Filters filters={filters} setFilters={setFilters} />
+        <Filters filters={filters} setFilters={setFilters} type="classes" />
         <ColumnSelector
           columns={translatedColumns}
           columnVisibility={columnVisibility}
